api_service/requirements.txt
# API Framework
fastapi>=0.110.0
uvicorn>=0.27.1
pydantic>=2.6.1
email-validator>=2.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.1

# Database
sqlalchemy>=2.0.27
psycopg2-binary>=2.9.9
alembic>=1.13.1

# AWS
boto3>=1.34.34

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Caching
redis>=5.0.1

# Logging
loguru>=0.7.2

# Utils
websockets>=12.0
tenacity>=8.2.3
jsonschema>=4.21.1
aiopg>=0.17.0
asyncpg>=0.29.0
aiohttp>=3.9.1
requests>=2.31.0
jwt>=1.3.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.5
httpx>=0.27.0
